pyproject.toml
[project]
name = "colony-strength-predictor"
version = "0.1.0"
description = "A machine learning training pipeline for classifying bee colony strength using sensor data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlflow>=3.1.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
    "httpx>=0.27.0",  # Required for FastAPI TestClient
]
